worst case contracts
