+++ dev/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "zkevm_dev"
+version = "0.1.0"
+edition = "2021"
+license = "MIT OR Apache-2.0"
+
+[dependencies]
+eth-types = { git = "https://github.com/pinkiebell/zkevm-circuits.git", branch = "zkevm-chain" }
+serde = { version = "1.0.136", features = ["derive"] }
+serde_json = { version = "1.0.78", features = ["preserve_order"] }
+
+[features]
+default = []
+autogen = []
