+++ dev/src/lib.rs
@@ -1,0 +1,2 @@
+pub mod bytecode;
+pub mod genesis;
